feat: Add firstName and lastName fields to User interface and update user-related routes
This commit adds the firstName and lastName fields to the User interface in the backend's database.ts file. These fields are now included in the User interface to store the first name and last name of a user.

In the admin.ts file, the /users and /users/csv routes are updated to include the firstName and lastName fields when retrieving user data. This change ensures that the user's first name and last name are included in the response.

In the auth.ts file, a new /name route is added to handle updating the first name and last name of a user. This route requires authentication and checks if the user is authorized before updating the user's information.

These changes enhance the functionality of the application by allowing users to provide their first name and last name during registration and updating their information later on.